pyproject.toml
[project]
name = "fastapi-prod-skeleton"
version = "0.1.0"
description = "FastAPI skeleton with production-grade API design patterns"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "structlog>=24.2.0",
  "sqlalchemy>=2.0.30",
  "python-jose[cryptography]>=3.3.0",
  "prometheus-client>=0.20.0",
  "anyio>=4.4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0",
]

[tool.ruff]
line-length = 100
select = ["E","F","I","UP","B"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
